Return add 2/6/2020

+++ database/migrations/2019_12_27_161651_create_sale_items_table.php
@@ -23,7 +23,7 @@ class CreateSaleItemsTable extends Migration
             $table->foreign('product_model_id')->references('id')
                 ->on('product_models')->onUpdate('cascade')->onDelete('cascade');
             $table->string('product_name');
-            $table->decimal('orderQuantity',25,2);
+            $table->integer('orderQuantity');
             $table->decimal('price',25,2);
             $table->decimal('totalPrice',25,2);
             $table->timestamps();
